ListItem having weird marginLeft as described on issue #1274
This was a known issue, making lists look weird because the border-bottom is always cut on the left, and when you use background on listitem and items you can see the background lacking on left also. For some reason it is now marginLeft instead of paddingLeft, and a List needs to fit the component device width not having margins, because most of times you use it 100% wide and want to put some border-bottom`s on it and expect the border to cover 100%. And moreover, a ListItem having paddingRight, and instead of having a paddingLeft also to mantain simmetry you see a marginLeft? this just does not make sense. Changing to paddingLeft solved all my problems and i believe that will solve many problems of the comunnity also.